Create finisher-badge route
